# Optimize block lookups?